+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Adds support for the `<pre>` element using the `Goatee.Preformatted()` class.
 - Adds support for the `<nav>` element using the `Goatee.Nav()` class.
 
+### Changes
+
+- Adds `addHeader()` and `addRow()` methods to the `Goatee.Table()` class.
+
 ## [0.4.1] - 2025-04-14
 
 ### Changes